refactor: extract formatTimeAgo to shared utils
